Changed parts vector from object reference to pointer.
The reason for this change is that the UI components are linked with a
weak pointer to the part objects. However if the vector of objects
reference is resized, all the reference become invalid because by default
the vector will copy all objects if new memory are allocated. Now I
changed the reference to pointer so that the obj pointer from UI still
valid after vector is resized. However, this needs active management of
pointers. We may need to bring in shared_ptr later on to simplify the
management of pointers.